refactor(worker): split job definitions into jobs/ subpackage
Extract job functions from the monolithic worker.py into domain modules:
- jobs/eval.py: run_eval
- jobs/maintenance.py: sync_component_sources, maintain_clickhouse
- jobs/catalog.py: generate_insight_report, batch_generate_insights,
  refresh_model_catalog

worker.py retains only arq WorkerSettings, startup/shutdown hooks, and
job registration imports. Each job module is independently importable
and testable without importing other job modules. No module in jobs/
contains a direct import from ee/.
