+++ observal-server/jobs/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: 2026 Hari Srinivasan <harisrini21@gmail.com>
+# SPDX-License-Identifier: AGPL-3.0-only
+
+"""Background job modules for the arq worker."""
